issue-4875: mved write fresh blocks logic
