+++ cloud/blockstore/libs/storage/partition_common/events_private.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cloud/blockstore/libs/diagnostics/profile_log.h>
 #include <cloud/blockstore/libs/kikimr/components.h>
 #include <cloud/blockstore/libs/kikimr/events.h>
 #include <cloud/blockstore/libs/storage/core/disk_counters.h>
@@ -23,6 +24,7 @@ namespace NCloud::NBlockStore::NStorage {
     xxx(TrimFreshLog,              __VA_ARGS__)                                \
     xxx(WriteBlob,                 __VA_ARGS__)                                \
     xxx(PatchBlob,                 __VA_ARGS__)                                \
+    xxx(AddFreshBlocks,            __VA_ARGS__)                                \
 // BLOCKSTORE_PARTITION_COMMON_REQUESTS_PRIVATE
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -47,10 +49,15 @@ struct TEvPartitionCommonPrivate
 
     struct TOperationCompleted
     {
+        NProto::TPartitionStats Stats;
+
         ui64 TotalCycles = 0;
         ui64 ExecCycles = 0;
 
         ui64 CommitId = 0;
+
+        TVector<TBlockRange64> AffectedRanges;
+        TVector<IProfileLog::TBlockInfo> AffectedBlockInfos;
     };
 
     //
@@ -364,6 +371,33 @@ struct TEvPartitionCommonPrivate
     };
 
     //
+    // AddFreshBlocks
+    //
+
+    struct TAddFreshBlocksRequest
+    {
+        ui64 CommitId;
+        ui64 BlobSize;
+        TVector<TBlockRange32> BlockRanges;
+        TVector<IWriteBlocksHandlerPtr> WriteHandlers;
+
+        TAddFreshBlocksRequest(
+                ui64 commitId,
+                ui64 blobSize,
+                TVector<TBlockRange32> blockRanges,
+                TVector<IWriteBlocksHandlerPtr> writeHandlers)
+            : CommitId(commitId)
+            , BlobSize(blobSize)
+            , BlockRanges(std::move(blockRanges))
+            , WriteHandlers(std::move(writeHandlers))
+        {}
+    };
+
+    struct TAddFreshBlocksResponse
+    {
+    };
+
+    //
     // Events declaration
     //
 
@@ -383,6 +417,8 @@ struct TEvPartitionCommonPrivate
         EvPartCountersCombined,
         EvPatchBlobCompleted,
         EvWriteBlobCompleted,
+        EvWriteFreshBlocksCompleted,
+        EvZeroFreshBlocksCompleted,
 
         EvEnd
     };
@@ -408,6 +444,11 @@ struct TEvPartitionCommonPrivate
 
     using TEvPatchBlobCompleted =
         TResponseEvent<TPatchBlobCompleted, EvPatchBlobCompleted>;
+
+    using TEvWriteFreshBlocksCompleted =
+        TResponseEvent<TOperationCompleted, EvWriteFreshBlocksCompleted>;
+    using TEvZeroFreshBlocksCompleted =
+        TResponseEvent<TOperationCompleted, EvZeroFreshBlocksCompleted>;
 };
 
 }   // namespace NCloud::NBlockStore::NStorage
