+++ cloud/blockstore/libs/storage/partition_common/fresh_blocks_companion.h
@@ -26,9 +26,11 @@ public:
 class TFreshBlocksCompanion
 {
 private:
+    const TStorageConfigPtr Config;
     const EStorageAccessMode StorageAccessMode;
     const NProto::TPartitionConfig PartitionConfig;
     const NKikimr::TTabletStorageInfoPtr TabletStorageInfo;
+    const IBlockDigestGeneratorPtr BlockDigestGenerator;
 
     IFreshBlocksCompanionClient& Client;
 
@@ -44,9 +46,11 @@ private:
 
 public:
     TFreshBlocksCompanion(
+        TStorageConfigPtr config,
         EStorageAccessMode storageAccessMode,
         NProto::TPartitionConfig partitionConfig,
         NKikimr::TTabletStorageInfo* tabletStorageInfo,
+        IBlockDigestGeneratorPtr blockDigestGenerator,
         IFreshBlocksCompanionClient& client,
         TPartitionChannelsState& channelsState,
         TPartitionFreshBlobState& freshBlobState,
@@ -61,8 +65,42 @@ public:
         const NActors::TActorContext& ctx,
         ui64 persistedTrimFreshLogToCommitId);
 
+    void WriteFreshBlocks(
+        const NActors::TActorContext& ctx,
+        TArrayRef<TRequestInBuffer<TWriteBufferRequestData>> requestsInBuffer,
+        ui64 commitId);
+
+    void WriteFreshBlocksCompleted(
+        const NActors::TActorContext& ctx,
+        const NProto::TError& error,
+        ui64 commitId,
+        ui64 blockCount,
+        NActors::TActorId actorId);
+
+    void ZeroFreshBlocks(
+        const NActors::TActorContext& ctx,
+        TRequestInfoPtr requestInfo,
+        TBlockRange32 writeRange,
+        ui64 commitId);
+
+    void ZeroFreshBlocksCompleted(
+        const NActors::TActorContext& ctx,
+        const NProto::TError& error,
+        ui64 commitId,
+        ui64 blockCount,
+        NActors::TActorId actorId);
+
+    void RebootOnCommitIdOverflow(
+        const NActors::TActorContext& ctx,
+        const TStringBuf& requestName);
+
+public:
     void HandleLoadFreshBlobsCompleted(
         const TEvPartitionCommonPrivate::TEvLoadFreshBlobsCompleted::TPtr& ev,
+        const NActors::TActorContext& ctx);
+
+    void HandleAddFreshBlocks(
+        const TEvPartitionCommonPrivate::TEvAddFreshBlocksRequest::TPtr& ev,
         const NActors::TActorContext& ctx);
 
 private:
